Set timeout for received bits
This allowes to abort receiving when line goes idle. Since we use an
explicit wait time now for all states, there is no need for a special
"infinite" wait anymore. We can also use zero wait time for the idle
state, so this is also no longer a special case.

Signed-off-by: Soeren Moch <smoch@web.de>